+++ src/main/java/subway/exception/DuplicateLineException.java
@@ -13,7 +13,7 @@ public class DuplicateLineException extends ApplicationException {
     }
 
     @Override
-    HttpStatus status() {
+    public HttpStatus status() {
         return CONFLICT;
     }
 }
