+++ src/main/java/subway/exception/DuplicateStationException.java
@@ -14,7 +14,7 @@ public class DuplicateStationException extends ApplicationException {
 
 
     @Override
-    HttpStatus status() {
+    public HttpStatus status() {
         return CONFLICT;
     }
 }
