+++ src/main/java/subway/exception/NotFoundLineException.java
@@ -18,7 +18,7 @@ public class NotFoundLineException extends ApplicationException {
     }
 
     @Override
-    HttpStatus status() {
+    public HttpStatus status() {
         return NOT_FOUND;
     }
 }
