+++ src/main/java/subway/exception/controller/ExceptionResponse.java
@@ -1,0 +1,13 @@
+package subway.exception.controller;
+
+public class ExceptionResponse {
+    private final String message;
+
+    public ExceptionResponse(final String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
